date-fns and lodash are added for date formatting and name capitalizing
