+++ index.html
@@ -13,5 +13,5 @@
     </div>
   </body>
 
-  <script type="module" src="main.js"></script>
+  <script src="./dist/main.js"></script>
 </html>
