+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "webdev-dom-homework",
+  "version": "1.0.0",
+  "description": "Comments",
+  "main": "main.js",
+  "scripts": {
+    "dev": "webpack --mode development --watch"
+  },
+  "author": "katiavasileva",
+  "license": "ISC",
+  "dependencies": {
+    "date-fns": "^3.6.0",
+    "lodash": "^4.17.21"
+  },
+  "devDependencies": {
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
