+++ webpack.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    entry: "./main.js",
+    output: {
+        filename: "main.js" 
+    }
+}
